nikolay-news/requirements.txt
requests>=2.25.1
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
mysql-connector-python>=8.0.0
python-multipart>=0.0.5
jinja2>=3.0.0
email-validator>=1.1.0
